test(.storybook): 스토리북 preview 초기 세팅
